# removed enumerator value support for EnumMap::getHash() 